HACK: bootm: Remove HAB authentication for now
As we want to use FIT images and check the signature with the
FIT_SIGNATURE way of things, remove the call to HAB authentication.

The proper way would be to add a new CONFIG_SPL_IMX_HAB, and have the
possibility to check the imx-boot FIT part with HAB only.

Signed-off-by: Jean-Michel Hautbois <jhautbois-ext@witekio.com>